+++ google2/CHANGELOG.md
@@ -4,13 +4,20 @@ This file documents changes to the `workbench-google2` library, including notes 
 
 ## 0.5
 
+Added
+- Add `getBlob`, `createObject`
+
 Changed
 - Use linebacker for blocking execution context
 - Moved `org.broadinstitute.dsde.workbench.google.GoogleKmsService` to `org.broadinstitute.dsde.workbench.google2.GoogleKmsService`
 - Add optional generation parameter to `removeObject`
 - Add insertBucket, which supports adding bucket labels
+- Deprecate `getObject`, `unsafeGetObject`, and add `getBlobBody`, `unsafeGetObjectBody`
+- provide `text/plain` as default `objectType` for `storeObject`
+- Bump `http4sVersion` to `0.20.3`
+- Deprecate `storeObject`, and add `createObject` that returns `Blob`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.5-ea3ee30"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.5-TRAVIS-REPLACE-ME"`
 
 ## 0.4